fix langs

+++ resources/views/account/languages.blade.php
@@ -106,24 +106,28 @@
                     </button>
                 {!! Form::close() !!}
             </div>
-            @foreach ($lineStats['on'] as $lang => $data)
-                <div class="popup__content popup__unavailable popup__del{{$data['langID']}}">
-                    <a href="#" class="popup__close">close</a>
-                    <h2 class="site__title site__title_center">Удаления языка {{$lang}}</h2>
-                    <div style="text-align:center">
-                        <a class="btn btn_8 btn_blue" href="/account/language/delete/{{$site->id}}/{{$data['langID']}}">Удалить</a>
+            @if (isset($lineStats['on']))
+                @foreach ($lineStats['on'] as $lang => $data)
+                    <div class="popup__content popup__unavailable popup__del{{$data['langID']}}">
+                        <a href="#" class="popup__close">close</a>
+                        <h2 class="site__title site__title_center">Удаления языка {{$lang}}</h2>
+                        <div style="text-align:center">
+                            <a class="btn btn_8 btn_blue" href="/account/language/delete/{{$site->id}}/{{$data['langID']}}">Удалить</a>
+                        </div>
                     </div>
-                </div>
-            @endforeach
-            @foreach ($lineStats['off'] as $lang => $data)
-                <div class="popup__content popup__unavailable popup__del{{$data['langID']}}">
-                    <a href="#" class="popup__close">close</a>
-                    <h2 class="site__title site__title_center">Удаления языка {{$lang}}</h2>
-                    <div style="text-align:center">
-                        <a class="btn btn_8 btn_blue" href="/account/language/delete/{{$site->id}}/{{$data['langID']}}">Удалить</a>
+                @endforeach
+            @endif
+            @if (isset($lineStats['off']))
+                @foreach ($lineStats['off'] as $lang => $data)
+                    <div class="popup__content popup__unavailable popup__del{{$data['langID']}}">
+                        <a href="#" class="popup__close">close</a>
+                        <h2 class="site__title site__title_center">Удаления языка {{$lang}}</h2>
+                        <div style="text-align:center">
+                            <a class="btn btn_8 btn_blue" href="/account/language/delete/{{$site->id}}/{{$data['langID']}}">Удалить</a>
+                        </div>
                     </div>
-                </div>
-            @endforeach
+                @endforeach
+            @endif
         </div>
     </div>
 @stop
